add new floating button to reset fields

+++ src/app/features/definitions/bonuses-types/presentation/bonuses-types.component.ts
@@ -1,5 +1,5 @@
-import {Component, OnInit} from '@angular/core';
-import {FormBuilder, Validators} from "@angular/forms";
+import { Component, OnInit } from '@angular/core';
+import { FormBuilder, Validators } from '@angular/forms';
 import { optionsBooleanGeneral } from 'src/app/core/core.interface';
 import { BonusesTypesFacade } from '../bonuses-types.facade';
 import { MessageType } from '../../../../shared/shared.interfaces';
@@ -16,32 +16,32 @@ export class BonusesTypesComponent implements OnInit {
   registerForm = this.fb.group({
     id: [''],
     name: ['', Validators.required],
-    isFamilyBonuse: ['', Validators.required],
+    isFamilyBonuse: ['', Validators.required]
   });
   constructor(
-      private fb: FormBuilder,
-      protected bonusesTypesFacade: BonusesTypesFacade,
-      protected sharedFacade: SharedFacade
+    private fb: FormBuilder,
+    protected bonusesTypesFacade: BonusesTypesFacade,
+    protected sharedFacade: SharedFacade
   ) {
     this.onSubmit();
   }
   ngOnInit() {
     this.edit = false;
   }
-  ngOnDestroy(): void {
-
-  }
+  ngOnDestroy(): void {}
   onSubmit(): void {
     this.registerForm.controls.id.setValue('');
     this.bonusesTypesFacade.GetBonusesType();
   }
   onDelete(Id: string): void {
-    this.edit = false;
-    this.bonusesTypesFacade.deleteBonusesType(Id);
-    this.registerForm.reset();
+    if (confirm('هل أنت متأكد من حذف هده العلاوة؟')) {
+      this.edit = false;
+      this.bonusesTypesFacade.deleteBonusesType(Id);
+      this.registerForm.reset();
+    }
   }
   activateBonusesType(bonuse): void {
-    this.bonusesTypesFacade.activateBonusesTypes(bonuse.id,!bonuse.isActive);
+    this.bonusesTypesFacade.activateBonusesTypes(bonuse.id, !bonuse.isActive);
     this.registerForm.reset();
   }
   onReset(): void {
@@ -51,20 +51,18 @@ export class BonusesTypesComponent implements OnInit {
   }
   onAdd(): void {
     if (this.registerForm.valid) {
-      if(this.edit) {
+      if (this.edit) {
         this.bonusesTypesFacade.UpdateBonusesType(this.registerForm?.value);
         this.onReset();
-      }else{
+      } else {
         this.bonusesTypesFacade.AddBonusesType(this.registerForm?.value);
         this.onReset();
-
       }
-    }else {
-      if(this.registerForm.value.name  == '' || this.registerForm.controls.name.invalid ){
+    } else {
+      if (this.registerForm.value.name == '' || this.registerForm.controls.name.invalid) {
         this.sharedFacade.showMessage(MessageType.warning, 'عفواً، الرجاء ادخال اسم العلاوة ', ['']);
         return;
-      }
-      else if( this.registerForm.controls.isFamilyBonuse.invalid ){
+      } else if (this.registerForm.controls.isFamilyBonuse.invalid) {
         this.sharedFacade.showMessage(MessageType.warning, 'عفواً، الرجاء ادخال علاوة عائلة', ['']);
         return;
       }
